Update pathname and location of form component imports
Adjustments have been made to several files to import the Form component from a new directory structure. It was previously imported directly from the 'components' folder, and is now properly located in the 'components/form' directory. This change is aligned with efforts to organise the project structure for easier navigation and management. It also includes the addition of a new component 'organization-selector-button' to manage organization selection for the user. Other minor refactoring of the code has also been made for improved efficiency and accuracy.