Update login and register URLs with redirect parameters
The href values of 'LinkButton' components for login and register actions in the 'layout.tsx' and 'page.tsx' files have been modified. Now, these URLs include a 'returnTo' parameter, improving the user's experience by redirecting them to a specific page (`/my` in this case) after they've logged in or registered.