Update button variant and add mouse move event to OrganizationCard
The button variants now include 'transition-colors'. For the OrganizationCard component, additional React hooks were introduced to capture and respond to the mouse movements. The tracking of mouse positions will be used for better interactive experiences in future. Also, modifications were made to the element's CSS class to add a hover effect.